tools/query-log-mining/requirements.txt
# Query Log Mining Toolkit Dependencies
# Install: pip install -r requirements.txt
# Or with uv: uv pip install -r requirements.txt

sqlglot>=20.0.0       # SQL parsing with Redshift dialect
polars>=0.20.0        # Fast DataFrame operations for large files
pyyaml>=6.0           # YAML output
openpyxl>=3.1.0       # Excel file reading
networkx>=3.0         # Graph algorithms (optional, for advanced analysis)
scikit-learn>=1.3     # Query clustering (optional, for Phase 2.5)
